engine/pyproject.toml
[tool.poetry]
name = "engine"
version = "0.1.0"
description = ""
authors = ["Wei-Cheng Pan <legnaleurc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Pillow = "^8.0.0"
aiohttp = "^3.4.0"
"wcpan.drive.core" = "^3.0.0"
"wcpan.drive.crypt" = "^3.0.0"
"wcpan.drive.google" = "^8.0.0"
"wcpan.logger" = "^1.2.0"
"wcpan.watchdog" = "^1.0.0"
wcpan-drive-cli = "^2.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
